feat: support the layer data
